egauge.webapi.device.capture: correct storing of samples with same timestamp
The Capture.result() method failed to update the "name" variable after
incrementing "ci" in the loop processing the samples.  This had the
effect that the samples were stored for the wrong channel whenever
there was more than one sample per timestamp.  It's rare to have
samples with the same timestamp on EG4xxx but very common on EG3xxx or
egauge2 devices.

Fixed by updating "name" after incrementing "ci".